Use `NativeUserPrincipalLookupService` and simplify `NativePosixFileAttributeView`
- do not automatically determine the name of a user or group principal
- `UserPrincipal` and `GroupPrincipal` have a well defined equality and hashCode
