+++ javalib/src/main/scala/scala/scalanative/nio/fs/attribute/NativeUserPrincipalLookupService.scala
@@ -12,14 +12,14 @@ import scala.scalanative.nio.fs.UnixException
 final case class NativeUserPrincipal(uid: stat.uid_t)(name: Option[String])
     extends UserPrincipal {
   override def getName = {
-    name.getOrElse(uid.toString)
+    name getOrElse NativeUserPrincipalLookupService.getUsername(uid)
   }
 }
 
 final case class NativeGroupPrincipal(gid: stat.gid_t)(name: Option[String])
     extends GroupPrincipal {
   override def getName: String = {
-    name.getOrElse(gid.toString)
+    name getOrElse NativeUserPrincipalLookupService.getGroupName(gid)
   }
 }
 
@@ -32,6 +32,38 @@ object NativeUserPrincipalLookupService extends UserPrincipalLookupService {
 
       NativeGroupPrincipal(gid)(Some(group))
     }
+
+  private[attribute] def getGroupName(gid: stat.gid_t): String = Zone {
+    implicit z =>
+      val buf = alloc[grp.group]
+
+      errno.errno = 0
+      val err = grp.getgrgid(gid, buf)
+
+      if (err == 0) {
+        fromCString(buf._1)
+      } else if (errno.errno == 0) {
+        gid.toString
+      } else {
+        throw UnixException("getgrgid", errno.errno)
+      }
+  }
+
+  private[attribute] def getUsername(uid: stat.uid_t): String = Zone {
+    implicit z =>
+      val buf = alloc[pwd.passwd]
+
+      errno.errno = 0
+      val err = pwd.getpwuid(uid, buf)
+
+      if (err == 0) {
+        fromCString(buf._1)
+      } else if (errno.errno == 0) {
+        uid.toString
+      } else {
+        throw UnixException("getpwuid", errno.errno)
+      }
+  }
 
   private def getGroup(name: CString)(
       implicit z: Zone): Option[Ptr[grp.group]] = {
